Move test for previews
